+++ .github/workflows/pythonpackage.yml
@@ -55,12 +55,12 @@ jobs:
         pip install -r requirements_dev.txt
     - name: Test with pytest
       run: |
-        export PARITY_ADDRESS=0x068ed00cf0441e4829d9784fcbe7b9e26d4bd8d0
-        export PARITY_PASSWORD=secret
-        export PARITY_KEYFILE=tests/resources/data/key_file_1.json
-        export PARITY_ADDRESS1=0x00bd138abd70e2f00903268f3db08f2d25677c9e
-        export PARITY_PASSWORD1=node0
-        export PARITY_KEYFILE1=tests/resources/data/key_file_2.json
+        export PARITY_ADDRESS=0x00bd138abd70e2f00903268f3db08f2d25677c9e
+        export PARITY_PASSWORD=node0
+        export PARITY_KEYFILE=tests/resources/data/key_file_2.json
+        export PARITY_ADDRESS1=0x068ed00cf0441e4829d9784fcbe7b9e26d4bd8d0
+        export PARITY_PASSWORD1=secret
+        export PARITY_KEYFILE1=tests/resources/data/key_file_1.json
 
         pip install pytest
         pytest -v
